% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdd.R
\name{get_hmdd_causality}
\alias{get_hmdd_causality}
\title{Retrieve causal miRNA-disease associations from HMDD3}
\usage{
get_hmdd_causality(entity)
}
\arguments{
\item{entity}{A string corresponding to a disease or a microRNA ID. If the latter,
there must be an "hsa-" prefix.}
}
\value{
A tibble containing the information queried.
}
\description{
Queries HMDD3 causal associations of a specified entity, which can be either a disease or a microRNA,
then retrieves a tibble with the information.
}
\examples{

get_hmdd_causality("sepsis")
get_hmdd_causality("hsa-mir-101")
}
