% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hmdd_content}
\alias{get_hmdd_content}
\title{HMDD content acquisition}
\usage{
get_hmdd_content(url, obj_xpath)
}
\arguments{
\item{url}{A string corresponding to the url where to acquire the information.}

\item{obj_xpath}{A string referring to the element xpath.}
}
\description{
Utility function to scrape HMDD pages and acquire relevant information,
primarily the href attribute corresponding to the downloadable table,
which is then read.
}
