% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdd.R
\name{get_hmdd_net}
\alias{get_hmdd_net}
\title{Retrieve miRNA-target networks from HMDD3}
\usage{
get_hmdd_net(entity)
}
\arguments{
\item{entity}{A string corresponding to a disease or a microRNA ID. If the latter,
there must be an "hsa-" prefix.}
}
\value{
An edgelist tibble containing the information queried.
}
\description{
Queries HMDD3 miRNA-target interactions for a specified entity,
which can be either a disease or a microRNA, then retrieves an
edgelist tibble with the information.
}
\examples{

get_hmdd_net("silicosis")
get_hmdd_net("hsa-mir-101")
}
