% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirdb.R
\name{get_mirdb_net}
\alias{get_mirdb_net}
\title{Retrieve miRNA-gene interaction from miRDB}
\usage{
get_mirdb_net(mirna)
}
\arguments{
\item{mirna}{A string corresponding to a microRNA ID, must be preceded by the 3-letter
species code, e.g. 'hsa' or 'mmu'.}
}
\description{
Queries miRDB for predicted miRNA-target interactions, returning a dataframe.
Supports same species miRDB supports, such as Homo sapiens, Mus musculus and
Rattus norvegicus. Each miRNA has to be preceeded by it's 3-letter species id.
}
\examples{

get_mirdb_net("mmu-mir-100-3p")
get_mirdb_net("hsa-miR-101-2-5p")
}
