% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirtarbase.R
\name{get_mirtar_disease}
\alias{get_mirtar_disease}
\title{Retrieve miRNA-disease information from mirTarBase}
\usage{
get_mirtar_disease(disease)
}
\arguments{
\item{disease}{A string corresponding to the disease you want to query.}
}
\value{
A tibble containing the miRNAs associated with the specified disease.
}
\description{
Queries mirTarBase for miRNAs associated with the specified disease, then outputs
a tibble with the specified information.
}
\examples{

get_mirtar_disease("carcinoma")
}
